requirements/base.txt
appnope==0.1.0
asgiref==3.4.1
attrs==19.3.0
backcall==0.1.0
black==23.9.1
bleach==6.0.0
certifi==2023.7.22
click==8.1.7
dataclasses==0.6
decorator==4.4.1
defusedxml==0.6.0
Django==3.2
django-environ==0.4.5
django-extensions==2.2.5
django-filter==2.2.0
django-model-utils==4.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-summernote==0.8.20.0
djangorestframework==3.13.1
entrypoints==0.3
factory-boy==2.12.0
Faker==4.1.0
future==0.18.2
gunicorn==20.0.4
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.10.1
jedi==0.15.2
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
packaging==23.1
pandocfilters==1.4.2
parso==0.5.2
pathspec==0.9.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.4.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psycopg2-binary==2.9.7
ptyprocess==0.6.0
pydantic==1.9.1
Pygments==2.5.2
pyrsistent==0.15.7
python-dateutil==2.8.1
python-environ==0.4.54
pytz==2019.3
pyzmq==25.1.1
qtconsole==4.6.0
Send2Trash==1.5.0
sentry-sdk==1.0.0
serializers==0.2.4
six==1.13.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.3
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.4
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.0.0
